refactor(personal-assistant): extract browser, screentime, scheduling, inbox, x domains
Phase B (batch 3). Breaks the browser<->screentime mutual cycle via dependency
injection (each takes the other's needed methods as typed deps instead of
importing each other). browser also injects its base helpers
(getBrowserSettingsInternal/isBrowserPaused/requireBrowserAvailableForActions/
buildBrowserCompanion/recordBrowserAudit/getWorkflowDefinition); scheduling
injects the 4 connector send methods; inbox rebuilds its Gmail/X DM sources from
injected deps instead of passing the composed `this`; x injects
recordXPostAudit (base) + resolvePrimaryChannelPolicy (reminders).

Plugin typechecks clean in refactor scope (0 errors in src/lifeops/); tests
unchanged (105 pass / 4 pre-existing fails). No behaviour change.
